Create README.md
Documentation Project.
---
What you need to have installed and how install if you haven't, the knowledge you need.
How to move to the clone folder project. The Two main folder, json-server to emulated a service and the Ionic App called conFusion.
How to run the mobile app in your main browser and test it.
---
Still working in week 4.